Add random scene and add final output
* Add random scene of spheres with random materials to render
* Increase integer capacity of the number of samples per pixel that can
  be sampled
* Update rendered image

+++ include/color.h
@@ -6,4 +6,4 @@
 using Color = Eigen::Vector3d;
 
 auto writeColor(std::ostream &rOut, Color rPixelColor) -> void;
-auto writeColor(std::ostream &rOut, Color rPixelColor, uint8_t rSamplesPerPixel) -> void;
+auto writeColor(std::ostream &rOut, Color rPixelColor, uint32_t rSamplesPerPixel) -> void;
